backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
psycopg2-binary==2.9.9
networkx==3.2.1
numpy==1.26.4
python-decouple==3.8

